high precision timer as replacement for QTime added

+++ image.h
@@ -19,7 +19,7 @@ public:
 public slots:
     void reset_view();
     void update_data(std::shared_ptr<mpFlow::numeric::Matrix<mpFlow::dtype::real>> data,
-        int time_elapsed);
+        double time_elapsed);
     void update_gl_buffer();
 
 protected:
